Change database schema
Instead of tracking the exact path of each artifact in the database,
record the filename, architecture, and if it's a debug package.  This is
necessary to allow a configuration prefix to customize directory
structure.  It also has the side benefit of only tracking noarch
packages once, instead of once for each architecture.

To indicate this is a new schema, the database filename has changed from
rpms.sqlite to index.1.db.  This filename will allow future schema
changes (index.2.db, index.3.db, etc).  No schema migration will take
place, just ask the user to delete the old database file and generate a
new one with the `index` command.